+++ src/utils/Settings.cpp
@@ -1,4 +1,5 @@
 #include "Settings.hpp"
+#include "../utils/SavePathSettingValue.hpp"
 
 int64_t Settings::getMaxSessionLength() {
     return Mod::get()->getSettingValue<int64_t>("session-length");
@@ -14,4 +15,12 @@ bool Settings::isPracticeZeroDeathsEnabled() {
 
 bool Settings::getPauseMenuEnabled(){
     return Mod::get()->getSettingValue<bool>("pause-menu");
+}
+
+bool Settings::getLateSaveEnabled(){
+    return Mod::get()->getSettingValue<bool>("late-save");
+}
+
+std::filesystem::path Settings::getSavePath(){
+    return std::filesystem::path{static_cast<SavePathSettingValue*>(Mod::get()->getSetting("save-path"))->getSavedPath()};
 }